refactor: move etablissement to catalogue
formerly on tco
