+++ server/tests/jobs/sources/catalogue-etablissements-test.js
@@ -3,12 +3,12 @@ const { dbCollection } = require("../../../src/common/db/mongodb");
 const { createSource } = require("../../../src/jobs/sources/sources");
 const { importEtablissements } = require("../../utils/testUtils");
 const collectSources = require("../../../src/jobs/collectSources");
-const { mockTcoApi } = require("../../utils/apiMocks");
+const { mockCatalogueApi } = require("../../utils/apiMocks");
 
 describe("catalogue-etablissements", () => {
   it("Vérifie qu'on peut collecter des informations relatives aux établissements du catalogue", async () => {
     await importEtablissements();
-    mockTcoApi((client, responses) => {
+    mockCatalogueApi((client, responses) => {
       client
         .get("/entity/etablissements.ndjson")
         .query(() => true)
